ai-service/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings

# LangChain and LLM
langchain
langchain-community
langchain-openai
langchain-text-splitters

# Vector Database (PostgreSQL + pgvector)
psycopg2-binary
pgvector
sqlalchemy

# Document parsing
pypdf
python-docx
